requirements.txt
streamlit>=1.0.0
tensorflow>=2.0.0
numpy>=1.19.0
Pillow>=9.0.0
protobuf>=3.12.0